requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
numpy==1.26.2
python-jose==3.3.0
passlib==1.7.4